receive/requirements.txt
flask==2.3.3
redis==5.0.1
python-dotenv==1.0.0
twilio==8.10.0
gunicorn==21.2.0
redis-om==0.2.1
